requirements.txt
# Core dependencies
torch==2.0.1
transformers==4.30.2
accelerate==0.21.0
peft==0.4.0
bitsandbytes==0.40.2
trl==0.4.7

# Data processing
datasets==2.13.1
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2

# Configuration and logging
pyyaml==6.0
wandb==0.15.5
tensorboard==2.13.0
tqdm==4.65.0

# Evaluation
nltk==3.8.1
rouge==1.0.1
matplotlib==3.7.1
seaborn==0.12.2

# Notebook support
jupyter==1.0.0
ipywidgets==8.0.6

# Math-specific
sympy==1.12
scipy==1.10.1 